重构: 拆分 plugin-command-expander.ts 为 plugin-expander 四件套 (PR#487 follow-up #6)
按 PR #487 review 第 #6 项原作者建议拆分原 854 行单文件。基于 codex
+ gpt 多轮 review 收敛到不留 facade 的"硬切"方案，避免给上游 main
增加旧聚合入口技术债 / 让未来代码继续从旧入口 import：

- src/plugin-expander-context.ts (新, 64 行)
    ExecutionMode + ExpandContext + makeExpandContext
    deps: 仅 path（不含 DATA_DIR）
- src/plugin-expander-sentinel.ts (新, 114 行)
    sentinel 类型常量 + read/write JSON helper + PersistExpansionFn 契约
    deps: 无（纯函数，零依赖）
- src/plugin-expander-store.ts (新, 31 行；替代旧 plugin-expansion-store.ts)
    仅保留 DB 级 persistPluginExpansion，import 改自 sentinel
- src/plugin-expander-core.ts (新, 719 行)
    主展开逻辑（slash 解析、frontmatter 渲染、inline bash 调度、
    placeholder 替换、batch helper）+ 私有 helper
    deps: ./plugin-expander-context + ./plugin-expander-sentinel；
    严格不 import store 避免间接加载 db.ts

删除：
- src/plugin-command-expander.ts (854 → 0)
- src/plugin-expansion-store.ts (30 → 0)

零行为差异：纯文件重组 + import path 切换，不动函数签名、不改算法。
所有 JSDoc 与内部注释跟着 symbol 一并搬走。运行时日志标识符字符串
'plugin-command-expander: ...' 刻意保留不动——operator/SRE 可能对其
建立了 grep alert，改名属于行为变更。

跨文件 type 用法严格使用 import type 关键字（ExpandContext / ExecutionMode /
PluginExpansionSentinel / PersistExpansionFn / PluginCommandIndexEntry /
Resolution / InlineExecResult 等），避免 TS 编译产生不必要 runtime
require。

依赖图（严格无环 + 加载面无 regression）：
  plugin-expander-context.ts  → 无内部依赖
  plugin-expander-sentinel.ts → 无内部依赖
  plugin-expander-store.ts    → sentinel + db
  plugin-expander-core.ts     → context + sentinel (不 import store)

私有 symbol 安全：4 个新文件按需具名 export，私有 helper（parseSlashHead /
findInlineMatches / renderExpandedPrompt / describeFailure / resolvePluginRoot
/ truncate / isInsideAnyRange / replacePlaceholdersAndSpliceInline /
applyPlaceholders / renderFrontmatterSummary / fencedRanges 等）不带
export 关键字。

Caller 改动：
- src/index.ts：4 个 import 切到新模块（含 type ExpandContext）
- src/web.ts：5 个 import 切到新模块（含 type ExpandContext）
- src/group-queue.ts：1 行 JSDoc 注释中的旧文件名引用同步更新
- 4 个 expander 测试文件：改 dynamic import path / static import path
  - tests/plugin-command-expander.test.ts → tests/plugin-expander-core.test.ts (git mv)
  - tests/plugin-expander-crash-safety-sentinel.test.ts: 拆 dynamic import
  - tests/plugin-expander-mixed-admin-batch.test.ts: 拆 dynamic import + 变量名
  - tests/plugin-expander-routing-bugs.test.ts: 改 static import path

验证：make typecheck 0 错；make test 577 passed (39 files) 与基线一致；
私有 symbol grep 仅 core.ts 自身命中；db 加载面 0 命中（context/sentinel/core）；
循环依赖单向；行数对账符合预期。
